requirements.txt
Flask==2.0.2
paho_mqtt==1.6.1
python-dotenv==0.19.2
requests==2.22.0
tb_rest_client==1.3.2
